+++ srunner/tools/route_parser.py
@@ -14,10 +14,19 @@ import xml.etree.ElementTree as ET
 import carla
 from agents.navigation.local_planner import RoadOption
 from srunner.scenarioconfigs.route_scenario_configuration import RouteScenarioConfiguration
+from srunner.scenarioconfigs.scenario_configuration import ScenarioConfiguration, ActorConfigurationData
 
 # TODO  check this threshold, it could be a bit larger but not so large that we cluster scenarios.
 TRIGGER_THRESHOLD = 2.0  # Threshold to say if a trigger position is new or repeated, works for matching positions
 TRIGGER_ANGLE_THRESHOLD = 10  # Threshold to say if two angles can be considering matching when matching transforms.
+
+
+def convert_dict_to_transform(scenario_dict):
+    """Convert a JSON dict to a CARLA transform"""
+    return carla.Transform(
+        carla.Location(float(scenario_dict['x']), float(scenario_dict['y']), float(scenario_dict['z'])),
+        carla.Rotation(roll=0.0, pitch=0.0, yaw=float(scenario_dict['yaw']))
+    )
 
 
 class RouteParser(object):
@@ -27,19 +36,19 @@ class RouteParser(object):
     """
 
     @staticmethod
-    def parse_annotations_file(annotation_filename):
+    def parse_scenario_file_to_dict(scenario_file):
         """
-        Return the annotations of which positions where the scenarios are going to happen.
-        :param annotation_filename: the filename for the anotations file
+        Parses and returns the scenario file into a dictionary
+        :param scenario_file: the filename for the scenario file
         :return:
         """
 
-        with open(annotation_filename, 'r', encoding='utf-8') as f:
-            annotation_dict = json.loads(f.read())
+        with open(scenario_file, 'r', encoding='utf-8') as f:
+            scenario_dict = json.loads(f.read())
 
         final_dict = {}
 
-        for town_dict in annotation_dict['available_scenarios']:
+        for town_dict in scenario_dict['available_scenarios']:
             final_dict.update(town_dict)
 
         return final_dict  # the file has a current maps name that is an one element vec
@@ -47,7 +56,7 @@ class RouteParser(object):
     @staticmethod
     def parse_routes_file(route_filename, scenario_file, single_route=None):
         """
-        Returns a list of route elements.
+        Returns a list of route configuration elements.
         :param route_filename: the path to a set of routes.
         :param single_route: If set, only this route shall be returned
         :return: List of dicts containing the waypoints, id and town of the routes
@@ -113,110 +122,87 @@ class RouteParser(object):
                     weather.fog_distance = float(weather_attrib.attrib['fog_distance'])
                 if 'fog_density' in weather_attrib.attrib:
                     weather.fog_density = float(weather_attrib.attrib['fog_density'])
+                if 'scattering_intensity' in weather_attrib.attrib:
+                    weather.scattering_intensity = float(weather_attrib.attrib['scattering_intensity'])
+                if 'mie_scattering_scale' in weather_attrib.attrib:
+                    weather.mie_scattering_scale = float(weather_attrib.attrib['mie_scattering_scale'])
+                if 'rayleigh_scattering_scale' in weather_attrib.attrib:
+                    weather.rayleigh_scattering_scale = float(weather_attrib.attrib['rayleigh_scattering_scale'])
 
         return weather
 
     @staticmethod
-    def check_trigger_position(new_trigger, existing_triggers):
+    def get_trigger_position(scenario_trigger, existing_triggers):
         """
         Check if this trigger position already exists or if it is a new one.
-        :param new_trigger:
-        :param existing_triggers:
+        :param scenario_trigger: position to be checked
+        :param existing_triggers: list with all the already found position
         :return:
         """
-
-        for trigger_id in existing_triggers.keys():
-            trigger = existing_triggers[trigger_id]
-            dx = trigger['x'] - new_trigger['x']
-            dy = trigger['y'] - new_trigger['y']
+        for trigger in existing_triggers:
+            dx = trigger.location.x - scenario_trigger.location.x
+            dy = trigger.location.y - scenario_trigger.location.y
             distance = math.sqrt(dx * dx + dy * dy)
 
-            dyaw = (trigger['yaw'] - new_trigger['yaw']) % 360
+            dyaw = (trigger.rotation.yaw - scenario_trigger.rotation.yaw) % 360
             if distance < TRIGGER_THRESHOLD \
                     and (dyaw < TRIGGER_ANGLE_THRESHOLD or dyaw > (360 - TRIGGER_ANGLE_THRESHOLD)):
-                return trigger_id
+                return trigger
 
         return None
 
     @staticmethod
-    def convert_waypoint_float(waypoint):
+    def match_trigger_to_route(trigger_transform, route):
         """
-        Convert waypoint values to float
+        Check if the scenario is affecting the route.
+        This is true if the trigger position is very close to any route point
         """
-        waypoint['x'] = float(waypoint['x'])
-        waypoint['y'] = float(waypoint['y'])
-        waypoint['z'] = float(waypoint['z'])
-        waypoint['yaw'] = float(waypoint['yaw'])
-
-    @staticmethod
-    def match_world_location_to_route(world_location, route_description):
-        """
-        We match this location to a given route.
-            world_location:
-            route_description:
-        """
-        def match_waypoints(waypoint1, wtransform):
-            """
-            Check if waypoint1 and wtransform are similar
-            """
-            dx = float(waypoint1['x']) - wtransform.location.x
-            dy = float(waypoint1['y']) - wtransform.location.y
-            dz = float(waypoint1['z']) - wtransform.location.z
+        def is_trigger_close(trigger_transform, route_transform):
+            """Check if the two transforms are similar"""
+            dx = trigger_transform.location.x - route_transform.location.x
+            dy = trigger_transform.location.y - route_transform.location.y
+            dz = trigger_transform.location.z - route_transform.location.z
             dpos = math.sqrt(dx * dx + dy * dy + dz * dz)
 
-            dyaw = (float(waypoint1['yaw']) - wtransform.rotation.yaw) % 360
+            dyaw = (float(trigger_transform.rotation.yaw) - route_transform.rotation.yaw) % 360
 
             return dpos < TRIGGER_THRESHOLD \
                 and (dyaw < TRIGGER_ANGLE_THRESHOLD or dyaw > (360 - TRIGGER_ANGLE_THRESHOLD))
 
-        match_position = 0
-        # TODO this function can be optimized to run on Log(N) time
-        for route_waypoint in route_description:
-            if match_waypoints(world_location, route_waypoint[0]):
-                return match_position
-            match_position += 1
+        for position, [route_transform, _] in enumerate(route):
+            if is_trigger_close(trigger_transform, route_transform):
+                return position
 
         return None
 
     @staticmethod
-    def get_scenario_type(scenario, match_position, trajectory):
+    def get_scenario_subtype(scenario, route):
         """
-        Some scenarios have different types depending on the route.
+        Some scenarios have subtypes depending on the route trajectory,
+        even being invalid if there isn't a valid one. As an example,
+        some scenarios need the route to turn in a specific direction,
+        and if this isn't the case, the scenario should not be considered valid.
+        This is currently only used for validity purposes.
+
         :param scenario: the scenario name
-        :param match_position: the matching position for the scenarion
-        :param trajectory: the route trajectory the ego is following
+        :param route: route starting at the triggering point of the scenario
         :return: tag representing this subtype
-
-        Also used to check which are not viable (Such as an scenario
-        that triggers when turning but the route doesnt')
-        WARNING: These tags are used at:
-            - VehicleTurningRoute
-            - SignalJunctionCrossingRoute
-        and changes to these tags will affect them
         """
 
-        def check_this_waypoint(tuple_wp_turn):
-            """
-            Decides whether or not the waypoint will define the scenario behavior
-            """
-            if RoadOption.LANEFOLLOW == tuple_wp_turn[1]:
-                return False
-            elif RoadOption.CHANGELANELEFT == tuple_wp_turn[1]:
-                return False
-            elif RoadOption.CHANGELANERIGHT == tuple_wp_turn[1]:
+        def is_junction_option(option):
+            """Whether or not an option is part of a junction"""
+            if option in (RoadOption.LANEFOLLOW, RoadOption.CHANGELANELEFT, RoadOption.CHANGELANERIGHT):
                 return False
             return True
 
-        # Unused tag for the rest of scenarios,
-        # can't be None as they are still valid scenarios
-        subtype = 'valid'
+        subtype = None
 
-        if scenario == 'Scenario4':
-            for tuple_wp_turn in trajectory[match_position:]:
-                if check_this_waypoint(tuple_wp_turn):
-                    if RoadOption.LEFT == tuple_wp_turn[1]:
+        if scenario == 'Scenario4':  # Only if the route turns
+            for _, option in route:
+                if is_junction_option(option):
+                    if option == RoadOption.LEFT:
                         subtype = 'S4left'
-                    elif RoadOption.RIGHT == tuple_wp_turn[1]:
+                    elif option == RoadOption.RIGHT:
                         subtype = 'S4right'
                     else:
                         subtype = None
@@ -224,98 +210,75 @@ class RouteParser(object):
                 subtype = None
 
         if scenario == 'Scenario7':
-            for tuple_wp_turn in trajectory[match_position:]:
-                if check_this_waypoint(tuple_wp_turn):
-                    if RoadOption.STRAIGHT == tuple_wp_turn[1]:
+            for _, option in route:
+                if is_junction_option(option):
+                    if RoadOption.STRAIGHT == option:
                         subtype = 'S7opposite'
-                    else:
-                        subtype = None
-                    break  # Avoid checking all of them
-                subtype = None
-
-        if scenario == 'Scenario8':
-            for tuple_wp_turn in trajectory[match_position:]:
-                if check_this_waypoint(tuple_wp_turn):
-                    if RoadOption.LEFT == tuple_wp_turn[1]:
+                    break
+        elif scenario == 'Scenario8':  # Only if the route turns left
+            for _, option in route:
+                if is_junction_option(option):
+                    if option == RoadOption.LEFT:
                         subtype = 'S8left'
-                    else:
-                        subtype = None
-                    break  # Avoid checking all of them
-                subtype = None
-
-        if scenario == 'Scenario9':
-            for tuple_wp_turn in trajectory[match_position:]:
-                if check_this_waypoint(tuple_wp_turn):
-                    if RoadOption.RIGHT == tuple_wp_turn[1]:
+                    break
+        elif scenario == 'Scenario9':  # Only if the route turns right
+            for _, option in route:
+                if is_junction_option(option):
+                    if option == RoadOption.RIGHT:
                         subtype = 'S9right'
-                    else:
-                        subtype = None
-                    break  # Avoid checking all of them
-                subtype = None
+                    break
+        else:
+            subtype = 'valid'
 
         return subtype
 
     @staticmethod
     def scan_route_for_scenarios(route_name, trajectory, world_annotations):
         """
-        Just returns a plain list of possible scenarios that can happen in this route by matching
-        the locations from the scenario into the route description
-
-        :return:  A list of scenario definitions with their correspondent parameters
+        Filters all the scenarios that are affecting the route.
+        Returns a dictionary where each item is a list of all the scenarios that are close to each other
         """
-
-        # the triggers dictionaries:
-        existent_triggers = {}
-        # We have a table of IDs and trigger positions associated
         possible_scenarios = {}
-
-        # Keep track of the trigger ids being added
-        latest_trigger_id = 0
 
         for town_name in world_annotations.keys():
             if town_name != route_name:
                 continue
 
-            scenarios = world_annotations[town_name]
-            for scenario in scenarios:  # For each existent scenario
-                if "scenario_type" not in scenario:
+            town_scenarios = world_annotations[town_name]
+            for scenario_data in town_scenarios:
+
+                if "scenario_type" not in scenario_data:
                     break
-                scenario_name = scenario["scenario_type"]
-                for event in scenario["available_event_configurations"]:
-                    waypoint = event['transform']  # trigger point of this scenario
-                    RouteParser.convert_waypoint_float(waypoint)
-                    # We match trigger point to the  route, now we need to check if the route affects
-                    match_position = RouteParser.match_world_location_to_route(
-                        waypoint, trajectory)
-                    if match_position is not None:
-                        # We match a location for this scenario, create a scenario object so this scenario
-                        # can be instantiated later
+                scenario_name = scenario_data["scenario_type"]
 
-                        if 'other_actors' in event:
-                            other_vehicles = event['other_actors']
-                        else:
-                            other_vehicles = None
-                        scenario_subtype = RouteParser.get_scenario_type(scenario_name, match_position,
-                                                                         trajectory)
-                        if scenario_subtype is None:
-                            continue
-                        scenario_description = {
-                            'name': scenario_name,
-                            'other_actors': other_vehicles,
-                            'trigger_position': waypoint,
-                            'scenario_type': scenario_subtype,  # some scenarios have route dependent configs
-                        }
+                for scenario in scenario_data["available_event_configurations"]:
+                    # Get the trigger point of the scenario
+                    trigger_point = convert_dict_to_transform(scenario.pop('transform'))
 
-                        trigger_id = RouteParser.check_trigger_position(waypoint, existent_triggers)
-                        if trigger_id is None:
-                            # This trigger does not exist create a new reference on existent triggers
-                            existent_triggers.update({latest_trigger_id: waypoint})
-                            # Update a reference for this trigger on the possible scenarios
-                            possible_scenarios.update({latest_trigger_id: []})
-                            trigger_id = latest_trigger_id
-                            # Increment the latest trigger
-                            latest_trigger_id += 1
+                    # Check if the route passes through the scenario
+                    match_position = RouteParser.match_trigger_to_route(trigger_point, trajectory)
+                    if match_position is None:
+                        continue
 
-                        possible_scenarios[trigger_id].append(scenario_description)
+                    # Check the route has the correct topology
+                    subtype = RouteParser.get_scenario_subtype(scenario_name, trajectory[match_position:])
+                    if subtype is None:
+                        continue
 
-        return possible_scenarios, existent_triggers
+                    # Parse the scenario data
+                    scenario_config = ScenarioConfiguration()
+                    scenario_config.type = scenario_name
+                    scenario_config.subtype = subtype
+                    scenario_config.trigger_points = [trigger_point]
+                    for other in scenario.pop('other_actors', []):
+                        scenario_config.other_actors.append(ActorConfigurationData.parse_from_dict(other, 'scenario'))
+                    scenario_config.other_parameters.update(scenario)
+
+                    # Check if its location overlaps with other scenarios
+                    existing_trigger = RouteParser.get_trigger_position(trigger_point, possible_scenarios.keys())
+                    if existing_trigger:
+                        possible_scenarios[existing_trigger].append(scenario_config)
+                    else:
+                        possible_scenarios.update({trigger_point: [scenario_config]})
+
+        return possible_scenarios